+++ src/App.tsx
@@ -64,6 +64,7 @@ import Menu from '@material-ui/core/Menu'
 import MenuItem from '@material-ui/core/MenuItem'
 import { faFileExcel } from '@fortawesome/free-solid-svg-icons'
 import { ListItemIcon, ListItemText } from '@material-ui/core'
+import OccupancyUpdater, { occupancyRates } from './utils/occupancy-updater'
 
 class App extends React.Component <{
 	classes: CreateCSSProperties
@@ -87,7 +88,8 @@ class App extends React.Component <{
 		darkMode: false,
 		multiShiftMode: false,
 		inhibitMultiShiftModeChange: false,
-		colorPicker: { show: false as boolean, course: undefined as (undefined | Course)  }
+		colorPicker: { show: false as boolean, course: undefined as (undefined | Course)  },
+		newDomainDialog: false
 	}
 	savedStateHandler: SavedStateHandler
 	selectedDegrees: Degree[] = []
@@ -100,6 +102,7 @@ class App extends React.Component <{
 	warningTitle = ''
 	warningContent = ''
 	warningContinue: () => void = () => {return}
+	newDomainURL = SavedStateHandler.DOMAIN
 
 	// eslint-disable-next-line
 	constructor(props: any) {
@@ -117,12 +120,17 @@ class App extends React.Component <{
 		this.onChangeDarkMode = this.onChangeDarkMode.bind(this)
 		this.onChangeMultiShiftMode = this.onChangeMultiShiftMode.bind(this)
 		this.exportToExcel = this.exportToExcel.bind(this)
+		this.updateShiftOccupancies = this.updateShiftOccupancies.bind(this)
 
 		this.chosenSchedule = React.createRef()
 		this.topBar = React.createRef()
 		this.colorPicker = React.createRef()
 
 		this.theme = this.getTheme(this.state.darkMode)
+
+		// Set occupancy updater
+		OccupancyUpdater.getInstance().changeRate(occupancyRates['Off'])
+		OccupancyUpdater.setUpdateFunction(this.updateShiftOccupancies)
 
 		this.savedStateHandler = new SavedStateHandler(API.getUrlParams())
 
@@ -150,9 +158,14 @@ class App extends React.Component <{
 		// Set warning with all notices
 		const isWarned = this.savedStateHandler.getWarning()
 		if (!isWarned) {
-			this.setWarningShiftDegrees()
+			this.setWarningDialog()
 			this.savedStateHandler.setWarning(true)
 		}
+		
+		// Warn about new domain
+		const isWarnedDomain = this.savedStateHandler.getNewDomain()
+		this.newDomainURL = await this.getSharingURL()
+		this.setState({newDomainDialog: !isWarnedDomain})
 	}
 
 	async onSelectedDegree(selectedDegree: Degree[]): Promise<void> {
@@ -374,12 +387,16 @@ class App extends React.Component <{
 		this.setState({ hasAlert: false })
 	}
 
-	async getLink(): Promise<void> {
+	async getSharingURL(): Promise<string> {
 		const shifts = shortenDescriptions(this.state.selectedShifts)
 		const degrees = getDegreesAcronyms(this.state.selectedShifts)
 		const isMultishift = this.state.multiShiftMode.toString()
 		const params = [`${SavedStateHandler.SHIFTS}=${shifts}`, `${SavedStateHandler.DEGREES}=${degrees}`, `${SavedStateHandler.IS_MULTISHIFT}=${isMultishift}`]
-		const shortLink = await API.getShortUrl(params)
+		return await SavedStateHandler.getAppURL(params)
+	}
+
+	async getLink(): Promise<void> {
+		const shortLink = await this.getSharingURL()
 		const el = document.createElement('textarea')
 		el.value = shortLink
 		el.setAttribute('readonly', '')
@@ -509,7 +526,7 @@ class App extends React.Component <{
 		this.setState({classesDialog: true, loading: false})
 	}
 
-	setWarningShiftDegrees(): void {
+	setWarningDialog(): void {
 		this.warningTitle = i18next.t('warning.title')
 		this.warningContent = (i18next.t('warning.content', {returnObjects: true}) as string[]).join('\n\n')
 		this.warningContinue = () => {return}
@@ -542,6 +559,44 @@ class App extends React.Component <{
 				saveMenuAnchor: null
 			})
 		}
+	}
+
+	async updateShiftOccupancies(): Promise<void> {
+		const shiftsById: Record<string, Shift> = {}
+		const coursesToBeFetched = new Set<Course>()
+		
+		// NOTICE: For now we update only the selected shifts
+		this.state.selectedShifts.forEach((s) => {
+			shiftsById[s.getStoredId()] = s
+			coursesToBeFetched.add(s.course)
+		})
+
+		const updatedShifts = await Promise.all(Array.from(coursesToBeFetched).map(async (c) => {
+			let newShifts: Shift[] | null | undefined = await API.getCourseSchedules(c)
+
+			newShifts = newShifts?.filter((s) => {
+				const toUpdateShift = shiftsById[s.getStoredId()]
+				if (toUpdateShift !== undefined) {
+					// FIXME: Remove this, just for testing
+					// s.occupation.current = Math.round(s.occupation.max * Math.random())
+					// --
+					toUpdateShift.updateOccupancy(s.occupation)
+				}
+
+				return toUpdateShift !== undefined
+			})
+
+			return newShifts
+		}))
+
+		// TODO: Maybe this can be moved to the previous cycle
+		const newUpdatedShifts = updatedShifts.flat().filter((s) => {
+			return s !== undefined
+		})
+
+		this.setState({
+			selectedShifts: newUpdatedShifts
+		})
 	}
 
 	render(): ReactNode {
@@ -785,7 +840,7 @@ class App extends React.Component <{
 									</Link>
 								</Tooltip>
 								<Tooltip title="Daniel Gonçalves">
-									<Link href="https://dagoncalves.me" target="_blank" onClick={() => {return}} color="inherit">
+									<Link href="https://dang.pt" target="_blank" onClick={() => {return}} color="inherit">
 										<IconButton size="small" title="Daniel Gonçalves" onClick={() => {return}}>
 											<Avatar alt="Daniel Goncalves" src={`${process.env.PUBLIC_URL}/img/daniel.png`} />
 										</IconButton>
@@ -827,6 +882,23 @@ class App extends React.Component <{
 								<div />
 								<Button onClick={() => {this.warningContinue(); this.setState({warningDialog: false})}} color="primary">{i18next.t('warning.actions.continue') as string}</Button>
 								{/* <Button onClick={() => {this.setState({warningDialog: false})}} color="primary">{i18next.t('warning.actions.back') as string}</Button> */}
+							</DialogActions>
+						</Dialog>
+						<Dialog maxWidth='sm' fullWidth open={this.state.newDomainDialog}>
+							<DialogTitle style={{alignSelf: 'center'}}>
+								{i18next.t('new-domain.title', {domain: SavedStateHandler.DOMAIN?.replaceAll('https://', '')})}
+							</DialogTitle>
+							<DialogContent style={{display: 'flex', flexDirection: 'column'}}>
+								<Box style={{whiteSpace: 'pre-line', alignSelf: 'center'}}>
+									{(i18next.t('new-domain.content', {returnObjects: true, domain: SavedStateHandler.DOMAIN?.replaceAll('https://', '')}) as string[]).join('\n\n')}
+								</Box>
+								<br/>
+								<Button variant='contained' style={{alignSelf: 'center'}} href={this.newDomainURL} color="primary">
+									{i18next.t('new-domain.actions.access') as string}
+								</Button>
+							</DialogContent>
+							<DialogActions>
+								<div />
 							</DialogActions>
 						</Dialog>
 						<Dialog open={this.state.changelogDialog}>